move scylladb to abstract for tx_indexer
