+++ database/Cargo.toml
@@ -13,6 +13,12 @@ prettytable-rs = { version = "0.10", optional = true }
 num-bigint = { version = "0.3", optional = true }
 num-traits = { version = "0.2.15", optional = true }
 scylla = { version = "0.9.0", optional = true }
+tokio = { version = "1.19.2", features = [
+    "sync",
+    "time",
+    "macros",
+    "rt-multi-thread",
+] }
 tracing = "0.1.34"
 uuid = { version = "1.3.0", optional = true }
 
@@ -23,6 +29,6 @@ near-crypto = "0.17.0"
 near-indexer-primitives = "0.17.0"
 
 [features]
-scylla_db_tracing = ["dep:prettytable-rs", "dep:uuid", "dep:scylla", "dep:num-bigint", "dep:num-traits"]
 scylla_db = ["dep:scylla", "dep:num-bigint", "dep:num-traits"]
+scylla_db_tracing = ["dep:prettytable-rs", "dep:uuid", "scylla_db"]
 account_access_keys = []
