+++ database/src/base.rs
@@ -60,7 +60,7 @@ pub trait RpcDbManager {
         &self,
         account_id: &near_primitives::types::AccountId,
         block_height: near_primitives::types::BlockHeight,
-    ) -> anyhow::Result<scylla::frame::response::result::Row>;
+    ) -> anyhow::Result<std::collections::HashMap<String, Vec<u8>>>;
 
     /// Returns the near_primitives::views::ReceiptView at the given receipt_id
     async fn get_receipt_by_id(
@@ -123,7 +123,7 @@ pub trait StateIndexerDbManager {
         &self,
         account_id: near_indexer_primitives::types::AccountId,
         block_height: u64,
-    ) -> anyhow::Result<scylla::frame::response::result::Row>;
+    ) -> anyhow::Result<std::collections::HashMap<String, Vec<u8>>>;
 
     #[cfg(feature = "account_access_keys")]
     async fn add_account_access_keys(
@@ -189,5 +189,65 @@ pub trait StateIndexerDbManager {
     ) -> anyhow::Result<()>;
 
     async fn update_meta(&self, indexer_id: &str, block_height: u64) -> anyhow::Result<()>;
+    async fn get_last_processed_block_height(&self, indexer_id: &str) -> anyhow::Result<u64>;
+}
+
+#[async_trait::async_trait]
+pub trait TxIndexerDbManager {
+    async fn add_transaction(
+        &self,
+        transaction: readnode_primitives::TransactionDetails,
+        block_height: u64,
+    ) -> anyhow::Result<()>;
+
+    async fn add_receipt(
+        &self,
+        receipt_id: &str,
+        parent_tx_hash: &str,
+        block_height: u64,
+        shard_id: u64,
+    ) -> anyhow::Result<()>;
+
+    async fn update_meta(&self, indexer_id: &str, block_height: u64) -> anyhow::Result<()>;
+
+    async fn cache_add_transaction(
+        &self,
+        transaction_details: readnode_primitives::CollectingTransactionDetails,
+    ) -> anyhow::Result<()>;
+
+    async fn cache_add_receipt(
+        &self,
+        transaction_key: readnode_primitives::TransactionKey,
+        indexer_execution_outcome_with_receipt: near_indexer_primitives::IndexerExecutionOutcomeWithReceipt,
+    ) -> anyhow::Result<()>;
+
+    async fn get_transactions_to_cache(
+        &self,
+        start_block_height: u64,
+        cache_restore_blocks_range: u64,
+        max_db_parallel_queries: i64,
+    ) -> anyhow::Result<
+        std::collections::HashMap<
+            readnode_primitives::TransactionKey,
+            readnode_primitives::CollectingTransactionDetails,
+        >,
+    >;
+
+    async fn get_transaction_by_receipt_id(
+        &self,
+        receipt_id: &str,
+    ) -> anyhow::Result<readnode_primitives::CollectingTransactionDetails>;
+
+    async fn get_receipts_in_cache(
+        &self,
+        transaction_key: &readnode_primitives::TransactionKey,
+    ) -> anyhow::Result<Vec<near_indexer_primitives::IndexerExecutionOutcomeWithReceipt>>;
+
+    async fn cache_delete_transaction(
+        &self,
+        transaction_hash: &str,
+        block_height: u64,
+    ) -> anyhow::Result<()>;
+
     async fn get_last_processed_block_height(&self, indexer_id: &str) -> anyhow::Result<u64>;
 }
