+++ database/src/lib.rs
@@ -1,9 +1,11 @@
 mod base;
 pub use crate::base::RpcDbManager;
 pub use crate::base::StateIndexerDbManager;
+pub use crate::base::TxIndexerDbManager;
 mod scylladb;
 pub use crate::scylladb::base::ScyllaStorageManager;
 
 pub mod primitives;
 pub mod rpc_server;
 pub mod state_indexer;
+pub mod tx_indexer;
