+++ database/src/scylladb/mod.rs
@@ -1,3 +1,4 @@
 pub mod base;
 pub mod rpc_server;
 pub mod state_indexer;
+pub mod tx_indexer;
