+++ state-indexer/Cargo.toml
@@ -9,21 +9,16 @@ actix-web = "4.2.1"
 aws-sdk-s3 = "0.23.0"
 aws-types = "0.53.0"
 anyhow = "1.0.70"
-async-trait = "0.1.66"
 borsh = "0.10.2"
 clap = { version = "3.1.6", features = ["color", "derive", "env"] }
 dotenv = "0.15.0"
 futures = "0.3.5"
 hex = "0.4.3"
-http = "0.2"
 humantime = "2.1.0"
-itertools = "0.10.0"
 lazy_static = "1.4.0"
 openssl-probe = "0.1.5"
 prometheus = "0.13.1"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1.0.55"
-tokio = { version = "1.1", features = [
+tokio = { version = "1.28.2", features = [
     "sync",
     "time",
     "macros",
@@ -56,6 +51,6 @@ near-lake-framework = "0.7.2"
 [features]
 default = ["scylla_db"]
 tracing-instrumentation = []
-scylla_db_tracing = ["database/scylla_db_tracing", "database/scylla_db"]
 scylla_db = ["database/scylla_db"]
+scylla_db_tracing = ["database/scylla_db_tracing", "scylla_db"]
 account_access_keys = ["database/account_access_keys"]
